+++ source/board/ST_NUCLEO-F042K6/ST_NUCLEO-F042K6-leds.cpp
@@ -1,0 +1,42 @@
+/**
+ * \file
+ * \brief Definition of leds for ST,NUCLEO-F042K6
+ *
+ * \author Copyright (C) 2014-2018 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ *
+ * \par License
+ * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
+ * distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * \warning
+ * Automatically generated file - do not edit!
+ */
+
+#include "distortos/board/leds.hpp"
+
+#if defined(CONFIG_BOARD_LEDS_ENABLE) && DISTORTOS_BOARD_LEDS_COUNT != 0
+
+#include "distortos/chip/ChipOutputPin.hpp"
+
+namespace distortos
+{
+
+namespace board
+{
+
+/*---------------------------------------------------------------------------------------------------------------------+
+| global objects
++---------------------------------------------------------------------------------------------------------------------*/
+
+chip::ChipOutputPin leds[ledsCount]
+{
+#ifdef CONFIG_BOARD_LEDS_LD3_ENABLE
+		chip::ChipOutputPin{chip::Pin::pb3, false, chip::PinOutputSpeed::low, chip::PinPull::none, false, false},
+#endif	// def CONFIG_BOARD_LEDS_LD3_ENABLE
+};
+
+}	// namespace board
+
+}	// namespace distortos
+
+#endif	// defined(CONFIG_BOARD_LEDS_ENABLE) && DISTORTOS_BOARD_LEDS_COUNT != 0