+++ source/board/ST_NUCLEO-F042K6/include/distortos/board/leds.hpp
@@ -1,0 +1,85 @@
+/**
+ * \file
+ * \brief Declaration of leds for ST,NUCLEO-F042K6
+ *
+ * \author Copyright (C) 2014-2018 Kamil Szczygiel http://www.distortec.com http://www.freddiechopin.info
+ *
+ * \par License
+ * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0. If a copy of the MPL was not
+ * distributed with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ *
+ * \warning
+ * Automatically generated file - do not edit!
+ */
+
+#ifndef SOURCE_BOARD_ST_NUCLEO_F042K6_INCLUDE_DISTORTOS_BOARD_LEDS_HPP_
+#define SOURCE_BOARD_ST_NUCLEO_F042K6_INCLUDE_DISTORTOS_BOARD_LEDS_HPP_
+
+#include "distortos/distortosConfiguration.h"
+
+#include <cstddef>
+
+#ifdef CONFIG_BOARD_LEDS_LD3_ENABLE
+#define DISTORTOS_BOARD_LEDS_LD3_ENABLED	1
+#else	// !def CONFIG_BOARD_LEDS_LD3_ENABLE
+#define DISTORTOS_BOARD_LEDS_LD3_ENABLED	0
+#endif	// !def CONFIG_BOARD_LEDS_LD3_ENABLE
+
+/// count of leds on the board
+#define DISTORTOS_BOARD_LEDS_COUNT	(DISTORTOS_BOARD_LEDS_LD3_ENABLED)
+
+#if defined(CONFIG_BOARD_LEDS_ENABLE) && DISTORTOS_BOARD_LEDS_COUNT != 0
+
+namespace distortos
+{
+
+namespace chip
+{
+
+class ChipOutputPin;
+
+}	// namespace chip
+
+namespace board
+{
+
+/// count of leds on the board
+constexpr size_t ledsCount {DISTORTOS_BOARD_LEDS_COUNT};
+
+/*---------------------------------------------------------------------------------------------------------------------+
+| indexes of leds
++---------------------------------------------------------------------------------------------------------------------*/
+
+enum
+{
+#ifdef CONFIG_BOARD_LEDS_LD3_ENABLE
+		/// index of Ld3 (Green) from leds
+		ledsLd3Index,
+#endif	// def CONFIG_BOARD_LEDS_LD3_ENABLE
+};
+
+/*---------------------------------------------------------------------------------------------------------------------+
+| alternative indexes of leds
++---------------------------------------------------------------------------------------------------------------------*/
+
+#ifdef CONFIG_BOARD_LEDS_LD3_ENABLE
+
+/// alternative index of Ld3 (Green) from leds
+constexpr size_t ledsGreenIndex {ledsLd3Index};
+
+#endif	// def CONFIG_BOARD_LEDS_LD3_ENABLE
+
+/*---------------------------------------------------------------------------------------------------------------------+
+| indexed access to leds objects
++---------------------------------------------------------------------------------------------------------------------*/
+
+/// array with all leds objects
+extern chip::ChipOutputPin leds[ledsCount];
+
+}	// namespace board
+
+}	// namespace distortos
+
+#endif	// defined(CONFIG_BOARD_LEDS_ENABLE) && DISTORTOS_BOARD_LEDS_COUNT != 0
+
+#endif	// SOURCE_BOARD_ST_NUCLEO_F042K6_INCLUDE_DISTORTOS_BOARD_LEDS_HPP_